can validate assets
