+++ generate-assets/Cargo.toml
@@ -14,3 +14,4 @@ edition = "2018"
 toml = "0.5"
 serde = { version = "1", features = [ "derive" ] }
 rand = "0.8"
+regex = "1"
