+++ generate-assets/generate_assets.sh
@@ -2,4 +2,4 @@
 
 git clone --branch bevy-asset https://github.com/bevyengine/awesome-bevy assets
 
-cargo run -- assets ../content
+cargo run --bin generate -- assets ../content
